requirements.txt
flask==3.0.0
openai-whisper
torch

